Move Ordering to datatypes to avoid name clashes with Eq
This is a platform dependent kapt bug :)
Either way it actually belongs there, so this is fine
