+++ arrow-core/src/main/kotlin/arrow/core/extensions/id.kt
@@ -9,6 +9,7 @@ import arrow.core.ForId
 import arrow.core.Id
 import arrow.core.IdOf
 import arrow.core.Ior
+import arrow.core.Ordering
 import arrow.core.Tuple2
 import arrow.core.extensions.id.monad.monad
 import arrow.core.fix
@@ -32,7 +33,6 @@ import arrow.typeclasses.MonadFx
 import arrow.typeclasses.MonadSyntax
 import arrow.typeclasses.Monoid
 import arrow.typeclasses.Order
-import arrow.typeclasses.Ordering
 import arrow.typeclasses.Repeat
 import arrow.typeclasses.Selective
 import arrow.typeclasses.Semialign
