+++ arrow-core/src/main/kotlin/arrow/core/extensions/option.kt
@@ -3,13 +3,17 @@
 package arrow.core.extensions
 
 import arrow.Kind
+import arrow.core.EQ
 import arrow.core.Either
 import arrow.core.Eval
 import arrow.core.ForOption
+import arrow.core.GT
 import arrow.core.Ior
+import arrow.core.LT
 import arrow.core.None
 import arrow.core.Option
 import arrow.core.OptionOf
+import arrow.core.Ordering
 import arrow.core.SequenceK
 import arrow.core.Some
 import arrow.core.Tuple2
@@ -29,15 +33,12 @@ import arrow.typeclasses.Applicative
 import arrow.typeclasses.ApplicativeError
 import arrow.typeclasses.Apply
 import arrow.typeclasses.Crosswalk
-import arrow.typeclasses.EQ
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
 import arrow.typeclasses.Foldable
 import arrow.typeclasses.Functor
 import arrow.typeclasses.FunctorFilter
-import arrow.typeclasses.GT
 import arrow.typeclasses.Hash
-import arrow.typeclasses.LT
 import arrow.typeclasses.Monad
 import arrow.typeclasses.MonadCombine
 import arrow.typeclasses.MonadError
@@ -49,7 +50,6 @@ import arrow.typeclasses.Monoid
 import arrow.typeclasses.MonoidK
 import arrow.typeclasses.Monoidal
 import arrow.typeclasses.Order
-import arrow.typeclasses.Ordering
 import arrow.typeclasses.Repeat
 import arrow.typeclasses.Selective
 import arrow.typeclasses.Semialign
