+++ arrow-core/src/main/kotlin/arrow/core/extensions/ordering.kt
@@ -1,12 +1,12 @@
 package arrow.core.extensions
 
+import arrow.core.EQ
+import arrow.core.Ordering
 import arrow.extension
-import arrow.typeclasses.EQ
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Hash
 import arrow.typeclasses.Monoid
 import arrow.typeclasses.Order
-import arrow.typeclasses.Ordering
 import arrow.typeclasses.Semigroup
 import arrow.typeclasses.Show
