+++ arrow-core/src/main/kotlin/arrow/core/extensions/string.kt
@@ -1,10 +1,10 @@
 package arrow.core.extensions
 
+import arrow.core.Ordering
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Hash
 import arrow.typeclasses.Monoid
 import arrow.typeclasses.Order
-import arrow.typeclasses.Ordering
 import arrow.typeclasses.Semigroup
 import arrow.typeclasses.Show
 
@@ -42,7 +42,9 @@ fun String.Companion.show(): Show<String> =
   object : StringShow {}
 
 interface StringOrder : Order<String> {
-  override fun String.compare(b: String): Ordering = Ordering.fromInt(this.compareTo(b))
+  override fun String.compare(b: String): Ordering =
+    Ordering.fromInt(this.compareTo(b))
+
   override fun String.compareTo(b: String): Int = this.compareTo(b)
 }
